requirements.txt
openai==1.43.1
pandas==2.2.3
numpy>=1.21
scipy>=1.7
python-dotenv==1.0.1
PyMuPDF==1.24.10
beautifulsoup4==4.11.2
httpx==0.27.2